server/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose==3.3.0
passlib==1.7.4
pydantic==2.5.2
email-validator==2.1.0
python-multipart==0.0.6
bcrypt==4.0.1
python-dotenv==1.0.0